able to see details owner of plant_pictures

+++ plant-django-app/plant_pictures/views.py
@@ -2,7 +2,10 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.exceptions import NotFound
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.exceptions import PermissionDenied
+
+
 
 from .models import Picture
 from .serializers.common import PictureSerializer
@@ -10,7 +13,7 @@ from .serializers.populated import PopulatedPictureSerializer
 
 
 class PictureListView(APIView):
-  permission_classes = (IsAuthenticatedOrReadOnly, )
+  permission_classes = (IsAuthenticated, )
 
   def get(self, _request):
     pictures = Picture.objects.all()
@@ -21,6 +24,7 @@ class PictureListView(APIView):
 
   def post(self, request):
     print('REQUEST')
+    request.data["owner"] = request.user.id 
     picture_to_add = PictureSerializer(data=request.data)
     if picture_to_add.is_valid():
             picture_to_add.save()
@@ -29,7 +33,7 @@ class PictureListView(APIView):
 
 
 class PictureDetailView(APIView):
-  permission_classes = (IsAuthenticatedOrReadOnly, )
+  permission_classes = (IsAuthenticated, )
 
   def get_picture(self, pk):
     try:
@@ -42,10 +46,15 @@ class PictureDetailView(APIView):
     serialized_picture = PopulatedPictureSerializer(picture)
     return Response(serialized_picture.data, status=status.HTTP_200_OK)
 
-  def delete(self, _request, pk):
-        picture_to_delete = self.get_picture(pk=pk)
-        picture_to_delete.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+  def delete(self,request, pk):
+    try:
+        picture_to_delete = self.objects.get(pk=pk)
+    except self.DoesNotExist:
+      raise NotFound()
+    if picture_to_delete.owner != request.user:
+      raise PermissionDenied()
+    picture_to_delete.delete()
+    return Response(status=status.HTTP_204_NO_CONTENT)
   
   def put(self, request, pk):
         picture_to_edit = self.get_picture(pk=pk)
